light js code partition
