+++ app/views/tubes/index.html.erb
@@ -5,9 +5,9 @@
 <div id="tubes">
 
   <search-bar id="search-bar" data-searchable=".tube" data-searchableparent=".seedbox"></search-bar>
-  <%= button_tag 'Show Critical', onclick: "window.criticalButton()"%>
-  <%= button_tag 'Generate CSV', onclick: "window.generateCsvButton()"%>
-  <%= button_tag 'Clear Search', onclick: "window.clearButton()"%>
+  <%= button_tag 'Show Critical', onclick: "window.criticalButton()" %>
+  <%= button_tag 'Generate CSV', onclick: "window.csvButton()" %>
+  <%= button_tag 'Clear Search', onclick: "window.clearButton()" %>
   <%= link_to "New tube", new_tube_path %>
 
   <% @grouped_tubes.each do |seedbox, tubes| %>
@@ -29,4 +29,5 @@
   <% end %>
 </div>
 <%= javascript_include_tag "search.js", type: "module" %>
-<%= javascript_include_tag "tube-search-buttons.js" %>
+<%= javascript_include_tag "tube-search-buttons.js", type: "module" %>
+<%= javascript_include_tag "generate-csv.js", type: "module"%>
